requirements.txt
fastapi
pyyaml
uvicorn
pillow
pillow-heif
